LayoutTests/webgpu/whlsl-buffer-length-expected.txt

PASS successfullyParsed is true

TEST COMPLETE
PASS resultUint32Array[0] is 1337
